replaced the movement queue with just two vec3s
