+++ Assets/Scenes/Scene.unity
@@ -238,7 +238,7 @@ Transform:
   - {fileID: 1319632929}
   - {fileID: 128778336}
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &269223113
 Prefab:
@@ -286,6 +286,10 @@ Prefab:
     - target: {fileID: 114383748892741250, guid: ebb0c0ec43d6f4633878d34f88500b14,
         type: 2}
       propertyPath: movementDirection.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
@@ -354,7 +358,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 50, y: -30, z: 0}
 --- !u!1001 &623204268
 Prefab:
@@ -435,7 +439,7 @@ Prefab:
       objectReference: {fileID: 0}
     - target: {fileID: 4465344283250088, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
       propertyPath: m_RootOrder
-      value: 8
+      value: 7
       objectReference: {fileID: 0}
     - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
       propertyPath: m_Name
@@ -444,6 +448,10 @@ Prefab:
     - target: {fileID: 114383748892741250, guid: ebb0c0ec43d6f4633878d34f88500b14,
         type: 2}
       propertyPath: movementDirection.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
@@ -495,6 +503,10 @@ Prefab:
     - target: {fileID: 114383748892741250, guid: ebb0c0ec43d6f4633878d34f88500b14,
         type: 2}
       propertyPath: movementDirection.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents:
@@ -597,7 +609,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 20.130001, y: -35.013, z: 0.006}
 --- !u!114 &1216079885
 MonoBehaviour:
@@ -657,6 +669,10 @@ Prefab:
     - target: {fileID: 114383748892741250, guid: ebb0c0ec43d6f4633878d34f88500b14,
         type: 2}
       propertyPath: movementDirection.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
@@ -827,7 +843,7 @@ Transform:
   m_LocalScale: {x: 48.713593, y: 48.713604, z: 48.713604}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1467506060
 GameObject:
@@ -1069,6 +1085,10 @@ Prefab:
       propertyPath: movementDirection.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
   m_IsPrefabParent: 0
@@ -1240,6 +1260,10 @@ Prefab:
       propertyPath: movementDirection.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents:
     - {fileID: 114349848293956462, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
   m_ParentPrefab: {fileID: 100100000, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
@@ -1281,7 +1305,11 @@ Prefab:
       objectReference: {fileID: 0}
     - target: {fileID: 4465344283250088, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
       propertyPath: m_RootOrder
-      value: 3
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 1065938242863216, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: ebb0c0ec43d6f4633878d34f88500b14, type: 2}
