+++ Assets/Scripts/Components/MovementComponent.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 [DisallowMultipleComponent]
 public class MovementComponent : MonoBehaviour {
 	
-	public Queue<Vector3> positionQueue = new Queue<Vector3>();
-	public Vector3 movementDirection = Vector3.forward;
+//	public Queue<Vector3> positionQueue = new Queue<Vector3>();
+	public Vector3 movementDirection;
+	public Vector3 nextMovement;
 }
