+++ Assets/Scripts/Systems/SnakePregnancySystem.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: af99c74eb3d964e9b8f3919b1911ebd6
-timeCreated: 1510748555
+guid: 53dc18ddcb361419f98c0ca276593597
+timeCreated: 1511757018
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
